Task 0

+++ 0x00-python_variable_annotations/0-add.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
-"""This Module contains an add function
+"""
+This Module contains an add function
 """
 
 
-def add(a: int, b: int) -> float:
+def add(a: float, b: float) -> float:
     """Returns the sum of a and b"""
     return a + b
